LayoutTests/imported/w3c/web-platform-tests/webmessaging/without-ports/020-expected.txt

Blocked access to external URL http://www1.localhost:8800/webmessaging/without-ports/020-1.html
CONSOLE MESSAGE: Unable to post message to http://www1.localhost:8800. Recipient has origin http://localhost:8800.



Harness Error (TIMEOUT), message = null

NOTRUN cross-origin test
